⚡️ Replace proxy-memoize with our own magic
This commit get rid of proxy-memoize and instead uses lodashs memoize with some custom twists for immerjs.
This is based on two assumptions:
1. We only have a few selectors we really need to memoize and
2. We only care for state changes after they ran through immerjs once.

Because of 2. you now have to be careful when using the selectors after modifying the state.
If your current state is different from the memoized state, but immerjs didn't notice yet,
you will get the memoized value. See the 'memoize returns old value if called within immerjs'
test in selectors.spec.ts for an example